--- src/main/java/com/kidmily/algoga_server/admin/settings/AdminJwtProvider.java
@@ -1,100 +1,0 @@
-package com.kidmily.algoga_server.admin.settings;
-
-import io.jsonwebtoken.JwtException;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
-import jakarta.annotation.PostConstruct;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-import javax.crypto.SecretKey;
-import java.nio.charset.StandardCharsets;
-import java.util.Date;
-
-@Component
-public class AdminJwtProvider {
-
-    @Value("${jwt.secret-key}")
-    private String secretKey;
-
-    @Value("${jwt.access-token-expiration}")
-    private long accessTokenExpiration;
-
-    @Value("${jwt.refresh-token-expiration}")
-    private long refreshTokenExpiration;
-
-    private SecretKey key;
-
-    @PostConstruct
-    protected void init() {
-        this.key = Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8));
-    }
-
-    public String createAccessToken(Long managerId, String loginId, String role) {
-        return Jwts.builder()
-                .subject(loginId)
-                .claim("type", "ADMIN") // 🌟 토큰 타입 명시
-                .claim("id", managerId)
-                .claim("role", role)
-                .issuedAt(new Date())
-                .expiration(new Date(System.currentTimeMillis() + accessTokenExpiration))
-                .signWith(key)
-                .compact();
-    }
-
-    // 🌟 토큰 타입 조회 메서드 추가
-    public String getType(String token) {
-        return Jwts.parser().verifyWith(key).build()
-                .parseSignedClaims(token).getPayload().get("type", String.class);
-    }
-
-    // (refreshToken은 로그인 연장용이므로 굳이 PK나 Role을 안 넣어도 됩니다)
-    public String createRefreshToken(String loginId) {
-        return Jwts.builder()
-                .subject(loginId)
-                .issuedAt(new Date())
-                .expiration(new Date(System.currentTimeMillis() + refreshTokenExpiration))
-                .signWith(key)
-                .compact();
-    }
-
-    // 🌟 2. 추가: 필터에서 사용할 getId 메서드 생성 (빨간 줄 해결!)
-    public Long getId(String token) {
-        return Jwts.parser()
-                .verifyWith(key)
-                .build()
-                .parseSignedClaims(token)
-                .getPayload()
-                .get("id", Long.class);
-    }
-
-    public String getSubject(String token) {
-        return Jwts.parser()
-                .verifyWith(key)
-                .build()
-                .parseSignedClaims(token)
-                .getPayload()
-                .getSubject();
-    }
-
-    public String getRole(String token) {
-        return Jwts.parser()
-                .verifyWith(key)
-                .build()
-                .parseSignedClaims(token)
-                .getPayload()
-                .get("role", String.class);
-    }
-
-    public boolean validateToken(String token) {
-        try {
-            Jwts.parser()
-                    .verifyWith(key)
-                    .build()
-                    .parseSignedClaims(token);
-            return true;
-        } catch (JwtException | IllegalArgumentException e) {
-            return false;
-        }
-    }
-}
