--- src/main/java/com/kidmily/algoga_server/user/settings/JwtProvider.java
@@ -1,77 +1,0 @@
-package com.kidmily.algoga_server.user.settings;
-
-import io.jsonwebtoken.JwtException;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
-import jakarta.annotation.PostConstruct;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-import javax.crypto.SecretKey;
-import java.nio.charset.StandardCharsets;
-import java.util.Date;
-
-@Component
-public class JwtProvider {
-
-    @Value("${jwt.secret-key}")
-    private String secretKey;
-
-    @Value("${jwt.access-token-expiration}")
-    private long accessTokenExpiration;
-
-    @Value("${jwt.refresh-token-expiration}")
-    private long refreshTokenExpiration;
-
-    private SecretKey key; // 0.12.x 부터는 Key 대신 SecretKey 사용 권장
-
-    @PostConstruct
-    protected void init() {
-        this.key = Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8));
-    }
-
-    public String createAccessToken(String email) {
-        return Jwts.builder()
-                .subject(email)
-                .claim("type", "USER") // 🌟 토큰 타입 명시
-                .issuedAt(new Date())
-                .expiration(new Date(System.currentTimeMillis() + accessTokenExpiration))
-                .signWith(key)
-                .compact();
-    }
-
-    public String getType(String token) {
-        return Jwts.parser().verifyWith(key).build()
-                .parseSignedClaims(token).getPayload().get("type", String.class);
-    }
-
-    public String createRefreshToken(String email) {
-        return Jwts.builder()
-                .subject(email)
-                .issuedAt(new Date())
-                .expiration(new Date(System.currentTimeMillis() + refreshTokenExpiration))
-                .signWith(key)
-                .compact();
-    }
-
-    public String getSubject(String token) {
-        return Jwts.parser() // parserBuilder() -> parser()
-                .verifyWith(key) // setSigningKey() -> verifyWith()
-                .build()
-                .parseSignedClaims(token) // parseClaimsJws() -> parseSignedClaims()
-                .getPayload() // getBody() -> getPayload()
-                .getSubject();
-    }
-
-    public boolean validateToken(String token) {
-        try {
-            Jwts.parser()
-                    .verifyWith(key)
-                    .build()
-                    .parseSignedClaims(token);
-            return true;
-        } catch (JwtException | IllegalArgumentException e) {
-            return false;
-        }
-    }
-}
