+++ src/main/resources/application.yaml
@@ -23,6 +23,7 @@ spring:
   config:
     import:
       - optional:classpath:/domains/banner.yml
+      - optional:classpath:/domains/userProfile.yml
   mail:
     host: smtp.gmail.com
     port: 587
